requirements.txt
boto3
cryptography
numpy
toolz
Jinja2
click
apache-airflow-providers-amazon
flower
openai==1.17.1
pyspark
astronomer-cosmos[openlineage]